// src/enums/RoleName.ts
export enum RoleName {
  doppelganger = 'doppelganger',
  werewolf = 'werewolf',
  minion = 'minion',
  mason = 'mason',
  seer = 'seer',
  robber = 'robber',
  troublemaker = 'troublemaker',
  drunk = 'drunk',
  insomniac = 'insomniac',
  villager = 'villager',
  hunter = 'hunter',
  tanner = 'tanner',
}